Add global labels to all timeseries.
Due to the structure of targetmanager/targetpool/target and who has visiblity
of which data, this is a bit ugly in places (like upon target add/replace,
getting each targets' existing labels, merging in the global ones, storing them
back).  But I doubt it can be done much nicer without completely refactoring
the retrieval code, which we should do at some point.

At least this approach is compatible with configuration reloading
(barring other things that need to be done for that).