Drop metric names after transformations.
After many transformations, it doesn't make sense to keep the metric
names, since the result of the transformation is no longer that metric.
This drops the metric name after such transformations and makes the web
UI deal well with missing metric names.

This depends on the current branch on the following things:

- prometheus/client_golang needs to be at
  https://github.com/prometheus/client_golang/commit/e237cf15c6967a6184e423a08154d9c6eecb330c
  in branch "julius/int-fingerprints" (to be merged with new storage)

- prometheus/promdash needs to be at
  https://github.com/prometheus/promdash/commit/dd7691c9c28c36767ae9559db8eab6d6715dffa8

Change-Id: Ib3c8cad8d647d9854e8c653c424b8c235ccc231d